+++ packages/variable-resolver/src/browser/variable-resolver-service.ts
@@ -30,8 +30,8 @@ export interface VariableResolveOptions {
     configurationSection?: string;
     commandIdVariables?: CommandIdVariables;
     configuration?: unknown;
-    // Return 'undefined' if not all variables were successfully resolved.
-    checkAllResolved?: boolean;
+    // Track `escape` from variable resolution e.g. from an interactive command
+    trackEscape?: string[]
 }
 
 /**
@@ -65,7 +65,7 @@ export class VariableResolverService {
     async resolve<T>(value: T, options: VariableResolveOptions = {}): Promise<T | undefined> {
         const context = new VariableResolverService.Context(this.variableRegistry, options);
         const resolved = await this.doResolve(value, context);
-        if (options.checkAllResolved && !context.allDefined()) {
+        if (context.hasEscaped()) {
             return undefined;
         }
         return resolved as any;
@@ -150,6 +150,10 @@ export namespace VariableResolverService {
             return true;
         }
 
+        hasEscaped(): boolean {
+            return this.options.trackEscape ? this.options.trackEscape.length > 0 : false;
+        }
+
         async resolve(name: string): Promise<void> {
             if (this.resolved.has(name)) {
                 return;
@@ -170,7 +174,8 @@ export namespace VariableResolverService {
                         argument,
                         this.options.configurationSection,
                         this.options.commandIdVariables,
-                        this.options.configuration
+                        this.options.configuration,
+                        this.options.trackEscape
                     ));
                 // eslint-disable-next-line no-null/no-null
                 const stringValue = value !== undefined && value !== null && JSONExt.isPrimitive(value as ReadonlyJSONValue) ? String(value) : undefined;